chore: move transformer and types to common
