+++ packages/common/src/transformers/index.ts
@@ -1,3 +1,4 @@
+export * from './BarChartDataTransformer';
 export * from './ResultTransformers';
 export * from './VizConfigTransformers';
 export * from './VizLibTransformers';
